+++ packages/api/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/api
 
+## 2.1.0-beta.18
+
+## 2.1.0-beta.17
+
 ## 2.1.0-beta.16
 
 ### Minor Changes