+++ packages/cli.common/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/cli.common
 
+## 0.25.0-beta.18
+
+## 0.25.0-beta.17
+
 ## 0.25.0-beta.16
 
 ### Minor Changes