+++ packages/client.unstable/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/client.unstable
 
+## 2.1.0-beta.18
+
+## 2.1.0-beta.17
+
 ## 2.1.0-beta.16
 
 ## 2.1.0-beta.15