+++ packages/create-app.template.tutorial-todo-app/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/create-app.template.tutorial-todo-app
 
+## 2.1.0-beta.18
+
+## 2.1.0-beta.17
+
 ## 2.1.0-beta.16
 
 ## 2.1.0-beta.15